EnrollForCourse Page updated + backend; Redirected when clicked on course in searchPage;Initializers adjusted

+++ src/main/java/com/Lernado/managers/CourseRepository.java
@@ -3,4 +3,4 @@ package com.Lernado.managers;
 import com.Lernado.model.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Long> {}
+public interface CourseRepository extends JpaRepository<Course, Integer> {}
